Align e2e orchestration project structure with bazel gazelle.
- BUILD files under the orchestration directory should not have
  contradictions with `bazel run //:gazelle`.
