+++ e2etests/go_repositories.bzl
@@ -1,6 +1,6 @@
 load("@bazel_gazelle//:deps.bzl", "go_repository")
 
-def go_dependencies():
+def repos():
     go_repository(
         name = "co_honnef_go_tools",
         importpath = "honnef.co/go/tools",
