+++ e2etests/orchestration/artifacts/def.bzl
@@ -12,12 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-load("@bazel_gazelle//:def.bzl", "gazelle")
-
-GO_PREFIX = "github.com/google/android-cuttlefish/e2etests"
-
-gazelle(
-    name = "gazelle",
-    external = "external",
-    prefix = GO_PREFIX,
-)
+def aosp_artifact(name, build_id, build_target, artifact_name, out_name):
+    native.genrule(
+        name = name,
+        outs = [out_name],
+        visibility = ["//visibility:public"],
+        cmd = "$(location :fetch_aosp_artifact) -b "+build_id+" -t "+build_target+" -a "+artifact_name+" -o $@",
+        tools = [":fetch_aosp_artifact"],
+    )
