+++ e2etests/orchestration/artifacts/fetch_aosp_artifact.sh
